requirements.txt
aiogram>=3.2.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
pyyaml>=6.0
alembic>=1.12.0
python-dotenv>=1.0.0